requirements.txt
django==3.0.7
djangorestframework==3.11.0
mysqlclient==1.4.6
django-mysql==3.7.1